refactor(pcg): lift HaybaGrammar PODs + pure grammar fns to public header (no behavior change)
Task 0.0 of the strangler-fig migration. Moves namespace HaybaGrammar (8 PODs +
14 pure fns + EvalGuards/ExpandGrammar/MatchProductions/ParseGrammar + ReadStr/NumAttr)
verbatim into Public/pcg/HaybaGrammarTypes.h (inline). Monolith includes it; .cpp 1288->768.
Parity: room 8v/12t, tunnel 116v/216t == baseline (byte-identical).
